requirements_web.txt
# ML core
torch>=2.0.0
pytorch-lightning>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
PyYAML>=6.0

# Web service
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# LLM (OpenAI Vision + Feedback)
openai>=1.0.0

# Object detection
ultralytics>=8.0.0

# Utilities
python-dotenv>=1.0.0
